backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djoser==2.2.2
django-filter==24.2
Pillow==10.3.0
drf-extra-fields==3.7.0
django-debug-toolbar==4.3.0